Set broad lower limits on parameter values

Sets wide lower limits on parameter values for each TPC model

Usage

get_lower_lims(x, y, model_name)

Arguments

x

vector of temperature values

y

vector of rate values

model_name

the name of the model being fitted

Value

Named list of lower limits given the data and model being fitted

Author

Daniel Padfield

Francis Windram