Kamykowski model for fitting thermal performance curves

Kamykowski model for fitting thermal performance curves

kamykowski_1985(temp, tmin, tmax, a, b, c)

Arguments

temp
temperature in degrees centigrade

tmin
low temperature (ºC) at which rates become negative

tmax
high temperature (ºC) at which rates become negative

a
parameter with no biological meaning

b
parameter with no biological meaning

c
parameter with no biological meaning

Details

Equation: $$rate= a \cdot \big( 1 - exp^{-b\cdot \big(temp-t_{min}\big)}\big) \cdot \big( 1-exp^{-c \cdot \big(t_{max}-temp\big)}\big)$$

Start values in get_start_vals are derived from the data or sensible values from the literature.

Limits in get_lower_lims and get_upper_lims are derived from the data or based extreme values that are unlikely to occur in ecological settings.

Note

Generally we found this model easy to fit.

References

Kamykowski, Daniel. A survey of protozoan laboratory temperature studies applied to marine dinoflagellate behaviour from a field perspective. Contributions in Marine Science. (1985).

Examples

# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'kamykowski_1985')
mod <- minpack.lm::nlsLM(rate~kamykowski_1985(temp = temp, tmin, tmax, a, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)

#> 
#> Formula: rate ~ kamykowski_1985(temp = temp, tmin, tmax, a, b, c)
#> 
#> Parameters:
#>       Estimate Std. Error t value Pr(>|t|)    
#> tmin 1.634e+01  3.699e+00   4.418  0.00309 ** 
#> tmax 4.881e+01  1.020e+00  47.861 4.55e-10 ***
#> a    2.106e+02  1.197e+05   0.002  0.99865    
#> b    4.010e-04  2.283e-01   0.002  0.99865    
#> c    1.207e-01  4.252e-01   0.284  0.78480    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.3716 on 7 degrees of freedom
#> 
#> Number of iterations till stop: 96 
#> Achieved convergence tolerance: 1.49e-08
#> Reason stopped: Number of calls to `fcn' has reached or exceeded `maxfev' == 600.
#> 

est_params(mod)

#>       rmax  topt  ctmin  ctmax         e       eh      q10
#> 1 1.327307 38.11 16.342 48.814 0.7173544 1.738695 2.478356
#>   thermal_safety_margin tolerance_range  skewness
#> 1                10.704          32.472 -1.021341