Lactin2 model for fitting thermal performance curves

Lactin2 model for fitting thermal performance curves

lactin2_1995(temp, a, b, tmax, delta_t)

Arguments

temp
temperature in degrees centigrade

a
constant that determines the steepness of the rising portion of the curve

b
constant that determines the height of the overall curve

tmax
the temperature at which the curve begins to decelerate beyond the optimum (ºC)

delta_t
thermal safety margin (ºC)

Details

Equation: $$rate= = exp^{a \cdot temp} - exp^{a \cdot t_{max} - \bigg(\frac{t_{max} - temp}{\delta _{t}}\bigg)} + b$$

Start values in get_start_vals are derived from the data or sensible values from the literature.

Limits in get_lower_lims and get_upper_lims are derived from the data or based extreme values that are unlikely to occur in ecological settings.

Note

Generally we found this model easy to fit.

References

Lactin, D.J., Holliday, N.J., Johnson, D.L. & Craigen, R. Improved rate models of temperature-dependent development by arthropods. Environmental Entomology 24, 69-75 (1995)

Examples

# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'lactin2_1995')
mod <- minpack.lm::nlsLM(rate~lactin2_1995(temp = temp, a, b, tmax, delta_t),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)

#> 
#> Formula: rate ~ lactin2_1995(temp = temp, a, b, tmax, delta_t)
#> 
#> Parameters:
#>         Estimate Std. Error t value Pr(>|t|)    
#> a        0.06598    0.06419   1.028    0.334    
#> b       -1.33646    2.11999  -0.630    0.546    
#> tmax    51.81245    5.02562  10.310 6.76e-06 ***
#> delta_t 11.91681    0.93024  12.810 1.30e-06 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.3378 on 8 degrees of freedom
#> 
#> Number of iterations to convergence: 15 
#> Achieved convergence tolerance: 1.49e-08
#> 

est_params(mod)

#>       rmax   topt  ctmin ctmax         e       eh      q10
#> 1 1.357256 38.406 15.591 48.74 0.7173544 1.738695 2.478356
#>   thermal_safety_margin tolerance_range  skewness
#> 1                10.334          33.149 -1.021341