Perform a parallelised quick tpc fit across models and curves

Performs a parallelised TPC fit using nls_multstart and map. This function tries to use a sensible default configuration, however if you need to use the more esoteric elements of nls_multstart then you may need to construct your own running script.

Usage

quickfit_tpc_multi(
  d,
  model_names,
  temp,
  trait,
  start_adjusts = 0,
  iter = 150,
  lhstype = "none",
  gridstart = FALSE,
  force = FALSE
)

Arguments

d

the data to fit a model to

model_names

a vector of model names to fit as strings

temp

the column name (as a string) containing the temperature data

trait

the column name (as a string) containing the temperature data

start_adjusts

any adjustments to make to the lower and upper starting bounds. If 0 < start_adjusts < 1, this will be interpreted as a proportion of the base starting values.

iter

number of combinations of starting parameters which will be tried (as in nls_multstart)

lhstype

method to use for Latin Hypercube Sampling using lhs (as in nls_multstart)

gridstart

whether to run a gridstart approach (interpreting iter as the number of samples to take across each parameter, so 3 will become c(3,3,3) for a 3-parameter model)

force

whether to force a gridstart even with very large numbers of iterations

Value

A tibble of model fits

Note

The parameters temp, trait, start_adjusts, iter, lhstype, gridstart, or force can be specified per-model by providing a vector of values of a length equal to the number of models to be fit.

Author

Francis Windram

Examples

if (FALSE) { # \dontrun{
data("chlorella_tpc")
d2 <- subset(chlorella_tpc, curve_id <= 60)

# Set up daemons for parallelism
mirai::daemons(2)

quickfit_tpc_multi(d2, c("briere1_1999", "briere2_1999"), "temp", "rate")

quickfit_tpc_multi(d2, c("briere1_1999", "briere2_1999"), "temp", "rate", start_adjusts = 10)

quickfit_tpc_multi(d2, c("briere1_1999", "briere2_1999"), "temp", "rate", iter = c(100, 150))

mirai::daemons(0)
} # }